Refactor command classes
