+++ src/Valleysoft.Dredge/Commands/Manifest/GetCommand.cs
@@ -1,0 +1,28 @@
+﻿using Newtonsoft.Json;
+using Valleysoft.DockerRegistryClient;
+using Valleysoft.DockerRegistryClient.Models;
+
+namespace Valleysoft.Dredge.Commands.Manifest;
+
+public class GetCommand : CommandWithOptions<GetOptions>
+{
+    public GetCommand(IDockerRegistryClientFactory dockerRegistryClientFactory)
+        : base("get", "Queries a manifest", dockerRegistryClientFactory)
+    {
+    }
+
+    protected override Task ExecuteAsync()
+    {
+        ImageName imageName = ImageName.Parse(Options.Image);
+        return CommandHelper.ExecuteCommandAsync(imageName.Registry, async () =>
+        {
+            using IDockerRegistryClient client = await DockerRegistryClientFactory.GetClientAsync(imageName.Registry);
+
+            ManifestInfo manifestInfo = await client.Manifests.GetAsync(imageName.Repo, (imageName.Tag ?? imageName.Digest)!);
+
+            string output = JsonConvert.SerializeObject(manifestInfo.Manifest, Formatting.Indented);
+
+            Console.Out.WriteLine(output);
+        });
+    }
+}
