+++ src/Valleysoft.Dredge/Commands/Repo/ListOptions.cs
@@ -1,0 +1,20 @@
+﻿using System.CommandLine;
+
+namespace Valleysoft.Dredge.Commands.Repo;
+
+public class ListOptions : OptionsBase
+{
+    private readonly Argument<string> registryArg;
+
+    public string Registry { get; set; } = string.Empty;
+
+    public ListOptions()
+    {
+        registryArg = Add(new Argument<string>("repo", "Name of the container registry"));
+    }
+
+    protected override void GetValues()
+    {
+        Registry = GetValue(registryArg);
+    }
+}
