+++ src/Valleysoft.Dredge/Commands/Settings/SettingsCommand.cs
@@ -1,0 +1,13 @@
+﻿using System.CommandLine;
+
+namespace Valleysoft.Dredge.Commands.Settings;
+
+internal class SettingsCommand : Command
+{
+    public SettingsCommand()
+        : base("settings", "Commands related to Dredge settings")
+    {
+        AddCommand(new OpenCommand());
+        AddCommand(new ClearCacheCommand());
+    }
+}
