+++ src/Valleysoft.Dredge/Commands/Tag/ListCommand.cs
@@ -1,0 +1,38 @@
+﻿using Newtonsoft.Json;
+using Valleysoft.DockerRegistryClient;
+using Valleysoft.DockerRegistryClient.Models;
+
+namespace Valleysoft.Dredge.Commands.Tag;
+
+public class ListCommand : CommandWithOptions<ListOptions>
+{
+    public ListCommand(IDockerRegistryClientFactory dockerRegistryClientFactory)
+        : base("list", "Lists the tag contained in the container repository", dockerRegistryClientFactory)
+    {
+    }
+
+    protected override Task ExecuteAsync()
+    {
+        ImageName imageName = ImageName.Parse(Options.Repo);
+        return CommandHelper.ExecuteCommandAsync(imageName.Registry, async () =>
+        {
+            using IDockerRegistryClient client = await DockerRegistryClientFactory.GetClientAsync(imageName.Registry);
+
+            List<string> tags = new();
+
+            Page<RepositoryTags> tagsPage = await client.Tags.GetAsync(imageName.Repo);
+            tags.AddRange(tagsPage.Value.Tags);
+            while (tagsPage.NextPageLink is not null)
+            {
+                tagsPage = await client.Tags.GetNextAsync(tagsPage.NextPageLink);
+                tags.AddRange(tagsPage.Value.Tags);
+            }
+
+            tags.Sort();
+
+            string output = JsonConvert.SerializeObject(tags, Formatting.Indented);
+
+            Console.Out.WriteLine(output);
+        });
+    }
+}
