+++ src/Valleysoft.Dredge/Program.cs
@@ -1,5 +1,10 @@
 ﻿using System.CommandLine;
 using Valleysoft.Dredge;
+using Valleysoft.Dredge.Commands.Image;
+using Valleysoft.Dredge.Commands.Manifest;
+using Valleysoft.Dredge.Commands.Repo;
+using Valleysoft.Dredge.Commands.Settings;
+using Valleysoft.Dredge.Commands.Tag;
 
 DockerRegistryClientFactory clientFactory = new();
 RootCommand rootCmd = new("CLI for executing commands on a container registry's HTTP API.")
