+++ src/Valleysoft.Dredge/RegistryHelper.cs
@@ -1,0 +1,6 @@
+﻿namespace Valleysoft.Dredge;
+
+public static class RegistryHelper
+{
+    public const string McrRegistry = "mcr.microsoft.com";
+}
